+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Content length header missing when signing request [#27](https://github.com/rokwire/core-auth-library-go/issues/27)
 ### Added
 - Provide permanent claim for external ID [#24](https://github.com/rokwire/core-auth-library-go/issues/24)
 - Add session ID claim [#32](https://github.com/rokwire/core-auth-library-go/issues/32)